feat(onboarding): Sign in done
