+++ src/App.tsx
@@ -179,7 +179,7 @@ import Settings from "./modules/Dashboard/modules/Settings/Settings";
 import { CampusStudentList, ConnectDiscord } from "./modules/Dashboard/modules";
 import Refer from "./modules/Dashboard/modules/Refer/Refer";
 import Thread from "./modules/Public/ThreadsCard/modules/Thread";
-import SignIn from "./modules/Common/Authentication/pages/Onboarding/signin";
+import SignIn from "./modules/Common/Authentication/pages/Onboarding/SignIn/signin";
 import RoleOne from "./modules/Common/Authentication/pages/Onboarding/RoleOne";
 import RoleFour from "./modules/Common/Authentication/pages/Onboarding/RoleFour";
 import LandingPage from "./modules/Public/LearningCircles/pages/LandingPage";
@@ -224,6 +224,10 @@ function App() {
         },
         {
             path: "/signin",
+            element: <SignIn />
+        },
+        {
+            path: "/signup",
             element: <SignIn />
         },
         {
